Bugfixes, clean-up and restructure.
Fixed the failure to load font file after some time.
Moved the queries to the server outside the drawing functions (as it
normally should be) so that now we can redraw as often as we want
without the http_get call failing after some time (probably the server
gives up on us due to spamming). The server query is currently done
every 2 minutes.
Moved most logging to stderr. Still not good but better.